Free interwar-1936 techs + Romanian core removed
The base license cost of any tech was lowered from 1 to 0. This will prevent early game boosting via licenses, some of which was already banned in the rules. This should also buff minors with early game reseach.

In addition, Romania loses their core on Transylvania to prevent it from being returned after Yugo/Greece are capped.

-Thrasymachus